Build on each push. Fixed Linux build.

+++ deps/serial/src/impl/unix.cc
@@ -455,7 +455,7 @@ Serial::SerialImpl::reconfigurePort ()
   // Compensate for the stopbits_one_point_five enum being equal to int 3,
   // and not 1.5.
   if (stopbits_ == stopbits_one_point_five) {
-    byte_time_ns_ += ((1.5 - stopbits_one_point_five) * bit_time_ns);
+    byte_time_ns_ += ((1.5f - (unsigned int)stopbits_one_point_five) * bit_time_ns);
   }
 }
